comments for PR 11. - business logic insite the /api/leaderboard is moved to server/game-service.ts - score variable was renamed to completedLevels - totaltime is passed without rounding off but the while displaying in the home page its rounded off
